Add UI for groups v2 invites.

+++ Signal/src/ViewControllers/ConversationView/Cells/MediaAlbumCellView.swift
@@ -1,5 +1,5 @@
 //
-//  Copyright (c) 2019 Open Whisper Systems. All rights reserved.
+//  Copyright (c) 2020 Open Whisper Systems. All rights reserved.
 //
 
 import Foundation
@@ -138,7 +138,7 @@ public class MediaAlbumCellView: UIStackView {
                 tintView.autoPinEdgesToSuperviewEdges()
 
                 let moreCount = max(1, items.count - MediaAlbumCellView.kMaxItems)
-                let moreCountText = OWSFormat.formatInt(Int32(moreCount))
+                let moreCountText = OWSFormat.formatInt(moreCount)
                 let moreText = String(format: NSLocalizedString("MEDIA_GALLERY_MORE_ITEMS_FORMAT",
                                                                 comment: "Format for the 'more items' indicator for media galleries. Embeds {{the number of additional items}}."), moreCountText)
                 let moreLabel = UILabel()
