+++ Signal/src/ViewControllers/ThreadSettings/ConversationSettingsViewController+Header.swift
@@ -138,7 +138,7 @@ extension ConversationSettingsViewController {
 
         let groupMembersText = String(format: NSLocalizedString("CONVERSATION_SETTINGS_GROUP_MEMBER_COUNT_FORMAT",
                                                                 comment: "Format for the 'group member count' indicator in conversation settings view. Embeds {the number of group members}."),
-                                      OWSFormat.formatInt(Int32(groupThread.groupModel.groupMembership.nonPendingMembers.count)))
+                                      OWSFormat.formatInt(groupThread.groupModel.groupMembership.nonPendingMembers.count))
         builder.addSubtitleLabel(text: groupMembersText,
                                  font: .ows_dynamicTypeBody)
